Improve performance by forcing rayon to schedule statically
- Rayon is a work-stealing library. It does many unnecessary syncs in
the scenario of kmeans, because every work item takes the exact same
amount of time to compute. So by callin with_min_len() on the parallel
iterators of rayon, it can be force to evenly split the work into n
work-packages, where n is the amount of hardware threads available.